updated the layout controller duration to be a float

+++ Scripts/Runtime/UI/Layout/LayoutTimeline.cs
@@ -16,7 +16,7 @@ namespace Gobblefish.Layout {
 
         // The duration for the loading to take.
         [SerializeField]
-        private int m_Duration;
+        private float m_Duration;
 
         // The position within the timeline. 
         [SerializeField]
